Mediawiki 1.42.1

See merge request archlinux/archwiki!81

+++ resources/src/mediawiki.editRecovery/storage.js
@@ -2,7 +2,9 @@
  * Common indexedDB-access methods, only for use by the ResourceLoader modules this directory.
  */
 
+const config = require( './config.json' );
 const dbName = mw.config.get( 'wgDBname' ) + '_editRecovery';
+const editRecoveryExpiry = config.EditRecoveryExpiry;
 const objectStoreName = 'unsaved-page-data';
 
 var db = null;
@@ -10,18 +12,19 @@ var db = null;
 // TODO: Document Promise objects as native promises, not jQuery ones.
 
 /**
+ * @ignore
  * @return {jQuery.Promise} Promise which resolves on success
  */
 function openDatabaseLocal() {
-	return new Promise( function ( resolve, reject ) {
-		const schemaNumber = 2;
+	return new Promise( ( resolve, reject ) => {
+		const schemaNumber = 3;
 		const openRequest = window.indexedDB.open( dbName, schemaNumber );
 		openRequest.addEventListener( 'upgradeneeded', upgradeDatabase );
-		openRequest.addEventListener( 'success', function ( event ) {
+		openRequest.addEventListener( 'success', ( event ) => {
 			db = event.target.result;
 			resolve();
 		} );
-		openRequest.addEventListener( 'error', function ( event ) {
+		openRequest.addEventListener( 'error', ( event ) => {
 			reject( 'EditRecovery error: ' + event.target.error );
 		} );
 	} );
@@ -35,9 +38,9 @@ function upgradeDatabase( versionChangeEvent ) {
 	const keyPathParts = [ 'pageName', 'section' ];
 	var objectStore;
 
-	if ( versionChangeEvent.oldVersion === 0 ) {
+	db = versionChangeEvent.target.result;
+	if ( !db.objectStoreNames.contains( objectStoreName ) ) {
 		// ObjectStore does not yet exist, create it.
-		db = versionChangeEvent.target.result;
 		objectStore = db.createObjectStore( objectStoreName, { keyPath: keyPathParts } );
 	} else {
 		// ObjectStore exists, but needs to be upgraded.
@@ -48,25 +51,29 @@ function upgradeDatabase( versionChangeEvent ) {
 	if ( !objectStore.indexNames.contains( 'pageName-section' ) ) {
 		objectStore.createIndex( 'pageName-section', keyPathParts, { unique: true } );
 	}
-	if ( !objectStore.indexNames.contains( 'expiryDate' ) ) {
-		objectStore.createIndex( 'expiryDate', 'expiryDate' );
+	if ( !objectStore.indexNames.contains( 'expiry' ) ) {
+		objectStore.createIndex( 'expiry', 'expiry' );
 	}
 
 	// Delete old indexes.
 	if ( objectStore.indexNames.contains( 'lastModified' ) ) {
 		objectStore.deleteIndex( 'lastModified' );
 	}
+	if ( objectStore.indexNames.contains( 'expiryDate' ) ) {
+		objectStore.deleteIndex( 'expiryDate' );
+	}
 }
 
 /**
  * Load data relating to a specific page and section
  *
+ * @ignore
  * @param {string} pageName The current page name (with underscores)
  * @param {string|null} section The section ID, or null if the whole page is being edited
  * @return {jQuery.Promise} Promise which resolves with the page data on success, or rejects with an error message.
  */
 function loadData( pageName, section ) {
-	return new Promise( function ( resolve, reject ) {
+	return new Promise( ( resolve, reject ) => {
 		if ( !db ) {
 			reject( 'DB not opened' );
 		}
@@ -75,8 +82,23 @@ function loadData( pageName, section ) {
 		const findExisting = transaction
 			.objectStore( objectStoreName )
 			.get( key );
-		findExisting.addEventListener( 'success', function () {
+		findExisting.addEventListener( 'success', () => {
 			resolve( findExisting.result );
+		} );
+	} );
+}
+
+function loadAllData() {
+	return new Promise( ( resolve, reject ) => {
+		if ( !db ) {
+			reject( 'DB not opened' );
+		}
+		const transaction = db.transaction( objectStoreName, 'readonly' );
+		const requestAll = transaction
+			.objectStore( objectStoreName )
+			.getAll();
+		requestAll.addEventListener( 'success', () => {
+			resolve( requestAll.result );
 		} );
 	} );
 }
@@ -84,13 +106,14 @@ function loadData( pageName, section ) {
 /**
  * Save data for a specific page and section
  *
+ * @ignore
  * @param {string} pageName The current page name (with underscores)
  * @param {string|null} section The section ID, or null if the whole page is being edited
  * @param {Object} pageData The page data to save
  * @return {jQuery.Promise} Promise which resolves on success, or rejects with an error message.
  */
 function saveData( pageName, section, pageData ) {
-	return new Promise( function ( resolve, reject ) {
+	return new Promise( ( resolve, reject ) => {
 		if ( !db ) {
 			reject( 'DB not opened' );
 		}
@@ -98,16 +121,16 @@ function saveData( pageName, section, pageData ) {
 		// Add indexed fields.
 		pageData.pageName = pageName;
 		pageData.section = section || '';
-		pageData.expiryDate = getExpiryDate( 30 );
+		pageData.expiry = getExpiryDate( editRecoveryExpiry );
 
 		const transaction = db.transaction( objectStoreName, 'readwrite' );
 		const objectStore = transaction.objectStore( objectStoreName );
 
 		const request = objectStore.put( pageData );
-		request.addEventListener( 'success', function ( event ) {
+		request.addEventListener( 'success', ( event ) => {
 			resolve( event );
 		} );
-		request.addEventListener( 'error', function ( event ) {
+		request.addEventListener( 'error', ( event ) => {
 			reject( 'Error saving data: ' + event.target.errorCode );
 		} );
 	} );
@@ -116,11 +139,13 @@ function saveData( pageName, section, pageData ) {
 /**
  * Delete data relating to a specific page
  *
+ * @ignore
  * @param {string} pageName The current page name (with underscores)
+ * @param {string|null} section The section ID, or null if the whole page is being edited
  * @return {jQuery.Promise} Promise which resolves on success, or rejects with an error message.
  */
-function deleteData( pageName ) {
-	return new Promise( function ( resolve, reject ) {
+function deleteData( pageName, section ) {
+	return new Promise( ( resolve, reject ) => {
 		if ( !db ) {
 			reject( 'DB not opened' );
 		}
@@ -130,74 +155,79 @@ function deleteData( pageName ) {
 
 		const request = objectStore.openCursor();
 
-		request.onsuccess = function ( event ) {
+		request.addEventListener( 'success', ( event ) => {
 			const cursor = event.target.result;
 			if ( cursor ) {
 				const key = cursor.key;
-				if ( key[ 0 ] === pageName ) {
+				const deleteSection = ( !section && key[ 1 ] === '' ) || section === key[ 1 ];
+				if ( key[ 0 ] === pageName && deleteSection ) {
 					objectStore.delete( key );
 				}
 				cursor.continue();
 			} else {
 				resolve();
 			}
-		};
+		} );
 
-		request.onerror = function () {
+		request.addEventListener( 'error', () => {
 			reject( 'Error opening cursor' );
-		};
+		} );
 	} );
 }
 
 /**
- * Returns the date diff days in the future
+ * Returns the date diff seconds in the future
  *
- * @param {number} diff Days in the future
+ * @ignore
+ * @param {number} diff Seconds in the future
  * @return {number} Timestamp of diff days in the future
  */
 function getExpiryDate( diff ) {
-	const today = new Date();
-	const diffDaysFuture = new Date( today.getFullYear(), today.getMonth(), today.getDate() + diff );
-	return diffDaysFuture.getTime() / 1000;
+	return ( Date.now() / 1000 ) + diff;
 }
 
 /**
  * Delete expired data
  *
+ * @ignore
  * @return {jQuery.Promise} Promise which resolves on success, or rejects with an error message.
  */
 function deleteExpiredData() {
-	return new Promise( function ( resolve, reject ) {
+	return new Promise( ( resolve, reject ) => {
 		if ( !db ) {
 			reject( 'DB not opened' );
 		}
 
 		const transaction = db.transaction( objectStoreName, 'readwrite' );
 		const objectStore = transaction.objectStore( objectStoreName );
-		const expiryDate = objectStore.index( 'expiryDate' );
-		const today = Date.now() / 1000;
+		const expiry = objectStore.index( 'expiry' );
+		const now = Date.now() / 1000;
 
-		const expired = expiryDate.getAll( IDBKeyRange.upperBound( today, true ) );
+		const expired = expiry.getAll( IDBKeyRange.upperBound( now, true ) );
 
-		expired.onsuccess = function ( event ) {
+		expired.addEventListener( 'success', ( event ) => {
 			const cursors = event.target.result;
-			if ( cursors ) {
-				cursors.forEach( function ( cursor ) {
-					deleteData( cursor.pageName );
+			if ( cursors.length > 0 ) {
+				const deletions = [];
+				cursors.forEach( ( cursor ) => {
+					deletions.push( deleteData( cursor.pageName, cursor.section ) );
 				} );
+				Promise.all( deletions ).then( resolve );
 			} else {
 				resolve();
 			}
-		};
+		} );
 
-		expired.onerror = function () {
+		expired.addEventListener( 'error', () => {
 			reject( 'Error getting filtered data' );
-		};
+		} );
 	} );
 }
 
 /**
  * Close database
+ *
+ * @ignore
  */
 function closeDatabase() {
 	if ( db ) {
@@ -207,9 +237,10 @@ function closeDatabase() {
 
 module.exports = {
 	openDatabase: openDatabaseLocal,
-	closeDatabase: closeDatabase,
-	loadData: loadData,
-	saveData: saveData,
-	deleteData: deleteData,
-	deleteExpiredData: deleteExpiredData
+	closeDatabase,
+	loadData,
+	loadAllData,
+	saveData,
+	deleteData,
+	deleteExpiredData
 };
